### 2020-03-02 #19 @pantale
Reproduction in Computational Mechanics
**Original article:** Olivier Pantalé, Parallelization of an object-oriented FEM dynamics code: influence of the strategies on the Speedup. (2005) Advances in Engineering Software, 36 (6). 361-373.

Paper n°38 for the 10 years reproduciblity challenge.

**PDF URL:**  [article.pdf](https://github.com/ReScience/submissions/files/4274211/article.pdf)

**Scientific domain:**  Computational Mechanics
**Programming language:**  C++ and C
