### 2020-03-02 #19 @rougier
@labarba Could you edit this submission ? You only need one reviewer and I think @khinsen is ready to review. I can help you with the publication procedure at the end.
