### 2020-03-16 #19 @khinsen
@labarba @pantale Here comes my review!

The topic of this reproducton attempt is not just the result of a computation performed 16 years ago, but also its performance characteristics. This makes it a particularly interesting contribution to the reproducibility challenge because performance, being a property of a physical system, lies at the borderline between reproduction and replication. The fact that the author did not re-run unmodified code from 2005, but a combination of later versions and code written specifically for this challenge adds to the borderline status of this work.

Given the ongoing confusion between reproducibility and replicability issues in many domains of science, I would like to encourage the author to make the specific nature of his contribution stand out more clearly.

The paper starts with a description of the historical evolution of the DynELA code, which is interesting because it highlights the impact of technological evolution on research. However, with the discussion of a "classical", "parallal", and "modern" version plus variants, it is not really clear how the code ultimately used in the reproduction differs from the code used in the original paper.

In Figure 2, the left-hand side represents the state before the impact, so I would expect it to be uniformly blue, rather than uniformly red which stands for maximal plastic strain. While the agreement can certainly be qualified as good, the numerical values in Figure 2 are slightly larger than in Figure 9 of the original paper. Can this tendency be explained?

Figure 3 is impossible to understand without a minimal description of the four parallelization strategies being compared. It would be helpful to include the method descriptions from page 368 of the original paper, if necessary with some rewording for avoiding copyright conflicts.

Finally, two minor remarks on the text:

1. Page 1: "the preferred current standard is C++"
Preferred by who? The FEM community? Certainly not scientific computing in general.

2. Page 2: "classical mathematics does not propose the notion of 4th-order tensor"
I don't know what "classical" mathematics is, but mathematics certainly does have this notion. The standard term is "tensor of rank 4".
