### 2020-03-19 #19 @pantale
@khinsen @labarba Hi,
I'm currently working on a revised version of the paper, hope to be able to send it by beginning of next week.

Kindly Regards

Olivier Pantalé
