### 2020-03-20 #19 @khinsen
@pantale @labarba The revised article looks good! I particularly appreciate the added explanations about the parallelization methods, which I find clearer than the explanations in the original article.

I had a look at the code and briefly considered a test installation. However, there are no instructions at all. The file `CMakeLists.txt` suggests that I need to use CMake, but since I haven't touched it for at least 15 years, I wouldn't know where to start. @pantale, could you provide some instructions? Or is this a hopeless task for anyone but you?
