### 2020-03-21 #19 @pantale
@khinsen
Hi,
I updated the README file of the ReDynELA repository to answer your questions. You can have a try on compiling the code. You can contact me directly at pantale@enit.fr for more instructions.

Olivier
