### 2020-03-23 pantale/ReDynELA#1 @khinsen
Build failure due to missing `yylex`
An installation attempt via the supplied instructions fails because of an undefined reference `yylex`:
```
[100%] Linking CXX executable ../../bin/DynELA_solve
ld: /gnu/store/6gvg7cvwbphx3zp4jmy2i4zc9xz7gszq-profile/lib/libfl.so: undefined reference to `yylex'
```

The full build log is attached:
[build.log](https://github.com/pantale/ReDynELA/files/4367993/build.log)
