### 2020-03-23 #19 @khinsen
@pantale Thanks for the instructions! Unfortunately they don't work for me, see https://github.com/pantale/ReDynELA/issues/1.
