### 2020-03-23 pantale/ReDynELA#1 @pantale
Hi,

Pehaps because I'm using the flex-old package instead of the flex one.
Mime is version : 2.5.4a-10ubuntu2

Olivier
