### 2020-03-23 #19 @khinsen
@pantale I forgot a legal detail: could you specify a license for the code? Somewhere in https://github.com/pantale/ReDynELA, ideally by adding a file called LICENSE. My understanding (noting that I am most definitely not a lawyer) is that without a license, nobody may use the code for anything. I'd then be almost in jail already for having tried to compile it ;-)
