### 2020-03-23 pantale/ReDynELA#1 @khinsen
Indeed, with flex 2.5 it works for me as well. Could you please add this requirement to the installation instructions?
