### 2020-03-23 4bd7331
Instructions for use with Guix
