### 2020-03-23 #19 @khinsen
Assuming you have no constraints imposed by your employer, I'd got for one of (1) MIT (2) GPL. The main difference is that an MIT license allows others to improve and re-distribute your code in various ways, including commercial, whereas the GPL requires all derived versions to be GPL-licensed as well.

There's a site providing guidance: https://choosealicense.com/
