### 2020-03-23 a0126ac
Add license to source files and updated Readme
