### 2020-03-23 #19 @pantale
Thanks a lot.

I have added the MIT license file to the repository.

Olivier
