### 2020-03-23 pantale/ReDynELA#1 @pantale
Instructions for the flex 2.5 have been added to the README.md file.
