### 2020-03-23 #19 @khinsen
@labarba I am happy with the current state of the article and the code!
