### 2020-03-26 #19 @pantale
Hi,
@khinsen @labarba What do I have to do now for continuing this process ?

Olivier
