### 2020-03-26 #19 @khinsen
@pantale Putting on my EIC's hat: right now, nothing. It's @labarba as the handling editor who has to decide if the paper is accepted or needs additional review or additional revision.
