### 2020-04-20 #19 @pantale
@rougier Hi,

Do I have to do this myself, or is it in the process of the publication ?
And concerning the proposed paper, what about the future ?

Kindly Regards,

Olivier Pantalé
