### 2020-05-03 #19 @rougier
@labarba Are oyu satisfied with the modification, can we accept the paper ?

@pantale Sorry for the delay. Yes, you have to save the code yourself (very easy to do) and software heritage will give you back an ID. This will need to be included in your metadata but the current template is not up to date (I need to fix it). Then @labarba (or me) will make a PR to your repor to modify  your metadata.
