### 2020-05-04 #19 @pantale
@rougier Hi,
The process of saving the code to software heritage is currently done...
I have different ID, depending on directory or revision, which one I have to use, I guess directory, so the following one:
https://archive.softwareheritage.org/swh:1:dir:e45b8b7efb002231f4081c429c4ca33664d0812d;origin=https://github.com/pantale/ReDynELA/

Olivier
