### 2020-05-12 #19 @pantale
> Ok, thanks. Now waiting for @labarba aknowledgement.

Ok, I'm also waiting for instructions...
Olivier
