### 2020-05-31 #19 @labarba
[article-3_annotatedLB.pdf](https://github.com/ReScience/submissions/files/4708017/article-3_annotatedLB.pdf)
