### 2020-05-31 #19 @labarba
Editorial review for ReScience, May 31, 2020
---

**Original paper:**
Pantalé, O. (2005). Parallelization of an object-oriented FEM dynamics code: influence of the strategies on the Speedup. Advances in Engineering Software, 36(6), 361-373. https://oatao.univ-toulouse.fr/5378/1/Pantale_5378.pdf

The original paper presents the shared-memory parallelization of the DynaELA code using OpenMP. DynaELA is an explicit finite-element-method solver for elasto-plastic deformation of materials under high-speed impact. The most computationally intensive part of the finite-element solver is the computation of the internal forces on each element of the mesh. Four multi-threading techniques for this part of the solver are discussed in the original paper, each more intricate than the other. They were tested on a benchmark (high-speed impact of a copper rod), reporting speedups and parallel efficiency on up to 8 CPU cores. The tests ran on an 8-way Intel-based Compaq SMP server from 1999 (Intel Pentium III Xeon 700MHz).

The simplest strategy is to add a parallel-for directive on the main loop. On 4 cores, speedup was 2.88 and efficiency 72%, but on 8 cores speedup _decreased_ and efficiency was 29%. The next two parallelization techniques performed similarly, with speedups of 6.45 and 7.88 on 8 cores (efficiencies of 80.7 and 98.5%). The fourth strategy included a load balancing operator (compute time of internal forces varies from element to element), and led to super-linear speedup, presumably due to efficient cache use.

The original paper also presents a typical application run for a dynamic-traction simulation of a cylindrical projectile impacting in a tubular target. The results from DynELA were compared to those using the commercial package Abaqus, and were in agreement. For this case, speedup results with the full DynELA are also shown (5.61x on 8 processors).

The “Reproducibility Challenge” work used an Intel-based Dell server with 24 cores (Intel Xeon E5-2650, 2.20GHz). Naturally, any speedup or parallel-efficiency observations on machines so many generations apart cannot be comparable. But looking at the general findings from a _replicability_ standpoint, can we say that the findings are consistent? The new results confirm that the first strategy for multi-threading quickly reaches a limit of speedup, and the fourth strategy gives the best performance—these findings replicate the original work.

The results of the application demo (dynamic-traction simulation) are shown as plastic strain contours. I left a comment in the PDF about this, too, as follows. The assessment of the results is inadequate. The statement is comparing the maximum value shown in the contourplot legend: 2.6e-2 in the 2005 paper, and 2.58e-2 here. First of all, I don't see a mention of how many processors were used in each case. For two parallel runs, you cannot expect bit-for-bit matching of numerical resutls, especially those arrived at after gather operations. These results are matching! You can look at other diagnostics: is the element with maximum strain at the same location in both cases? What about the total length after deformation?

Congratulations to the author on the success at finding the code (which had not been archived properly), painstakingly merging with a “beta” version with OpenMP, and getting it to compile and run. This is a feat to be commended.

But I found the paper a slog to read, with some incorrect English usage and too verbose. In the end, I could not help myself and fully annotated the PDF with copy editing.

**The source code**
The GitHub repository has a LICENSE file with MIT License, but the source files have text for GNU. See for example: https://github.com/pantale/ReDynELA/blob/master/Sources/Applications/DynELA_solve/main.C
