### 2020-05-31 #19 @labarba
Note my comment in the PDF about the use of the word "validate." This is a technical word in this field that means comparing the results from a simulation with a physical experiment. The word should not be used with any different meaning.
