### 2020-05-31 #19 @labarba
Please note the annotations on the PDF were done in Acrobat and I'm not sure if any PDF viewer shows them.
