### 2020-06-01 #19 @labarba
An additional comment about the categorization of this paper as a Replication, or Reproduction.

It's tricky because in this case, it is both! The comparison of multi-threading strategies with OpenMP can be called a replication: we have different computing systems (several generations away from each other), and look at whether the experiments with the copper-rod impact test give the same findings. And they do: method 1 saturates and does not scale, and method 4 scales the best with increasing number of processors. (Note: use 'processors' and not 'CPU' in this context.)

The application demo (dynamic-traction simulation) is a reproduction: same code, same computational problem set-up, rerun the simulation and observe the same quantities. The numerical results match, and the reproduction is achieved.

I also want to comment that even if the reproduction was achieved, it is slightly misleading to call the original results 'reproducible.' They were not reproducible to anyone other than the original author, because code and data were not public. And even that was a challenge since the code was lost (but found) in old hard drives! The work is reproducible _now_, with the code made available with this work.

I think it is worthwhile to clarify this in the paper.
