### 2020-06-01 #19 @khinsen
@labarba Many submissions to the ten-year-challenge end up being at the borderline between reproduction and replicability. That's one of the main conclusions of the challenge, but not one that we had foreseen when writing the author guidelines. So if all contributions to the challenge end up with the "reproduction" label, that's our fault, not the authors'.

However, as you suggest, it's a good idea for authors to elaborate on this in their submissions.
