### 2020-06-02 #19 @pantale
@labarba @khinsen
Thanks a lot for the feedback.
I'm working on the corrections to the article.
Olivier
