### 2020-06-02 #19 @labarba
> Many submissions to the ten-year-challenge end up being at the borderline between reproduction and replicability.

Saying "borderline" is confusing. This submission contains BOTH replication (applies to the analysis of multi-threading strategies), and reproduction (applies to the benchmark on dynamic-traction simulation).

I want to add to my points about the dynamic-traction test. The author should remove speculations about whether he "rounded off" 2.58 into 2.6. The results are shown as a contour plot, and the maximum number shown in the legend was output by the plotting software (and unlikely to have been manipulated by hand). From my reading of the paper, I understand that a new or modified post-processing software was used. Thus, it's impossible to know where the difference comes from. One plotting routine may be choosing the maximum value of strain at a node of the element, and the other could be averaging the nodes, for example. Or (as I mentioned), it could simply be accumulation of floating-point errors in the gather operation on these two different multi-threaded runs. In sum, both results (2.58 and 2.6) should be considered equally credible approximations to the true value, and thus consistent.
