### 2020-06-02 #19 @pantale
@labarba @khinsen
All correction are done confirming to the annoted PDF. Only the clarification about reproduction/replication has to be done.
What about the license ? Do you want me to remove allusions to gnu license in source files ?
