### 2020-06-02 #19 @labarba
Re. license --- you have to pick one. Either remove all the text on GNU license from source files and leave the BSD-3 LICENSE file or replace the LICENSE file to match GNU. The difference, as Konrad explained, is BSD is more permissive and GNU is copyleft.

See if this helps:
https://figshare.com/articles/A_short_lecture_on_Open_Licensing/4516892/1
