### 2020-06-03 273db43
Remove old license text from sources
