### 2020-06-03 #19 @pantale
@labarba @khinsen
Hi,
Here is a revised version of the paper.
[article.pdf](https://github.com/ReScience/submissions/files/4721821/article.pdf)

As far as the changes are concerned, here is a short summary:
- Regarding the difference 2.58, 2.60, I updated the text and added the fact that we're comparing a 32-bit processor and a 64-bit processor on an explicit integration scheme, known to be sensitive to round-error calculations.
- I have added a small paragraph in the concluding section about the fact that I have both replication and replication, depending on the results or performance of parallel computing. I hope this will clarify the position of this paper in the challenge.
- All changes to the text have been taken into account and I thank you for your helpful comments.

I have removed all the text on the GNU license from the source files and made a new commit of the source files to github. The software heritage has also been updated.

Olivier
