### 2020-06-05 #19 @labarba
I'm satisfied with the changes. A few typos remain, including:

p.4, line 6: softwares >> software (an uncountable noun)
p.5, line 8: requested >> required
p.5, line 10: have >> has (subject: "one")
p.5, last paragraph: reproduce>>replicate
p.10, last line: repoducible>>reproducible
