### 2020-06-05 #19 @pantale
@labarba
Hi,
Thank you very much for these last corrections which I have taken into account in the revised version. I also have checked and corrected some other typos in the text.
**PDF**: [article.pdf](https://github.com/ReScience/submissions/files/4736050/article.pdf)
**GitHub paper source**: https://github.com/pantale/ReScience-paper2020

Olivier
