### 2020-06-15 #19 @rougier
@larbarba If satisiied with the corrections ping me if you need help with the publication process
