### 2020-06-15 #19 @labarba
Thanks, @rougier --- after the last small edits, the paper is ready to publish. I'm not up to speed with the mechanics of the acceptance workflow in ReScience. Appreciate if you can take it from here.
