### 2020-06-15 #19 @rougier
Sure. The semi-automated procedure is explained here: https://github.com/ReScience/articles

You'll need to clone this repository and have the `article.pdf` and `'metadata.yaml` at the root directory. The metadata need first to be filled by you or @pantale and the pdf regenerated. Then you'll need to reserve a DOI for the article such that it can be regenerated with proper DOI. Best is to try first on the sandbox Zenodo server because things can go messy.

Last option (ans easier for you) is I publish the paper :)
