### 2020-06-15 #19 @rougier
@labarba Thanks for your work (you can unsubscribe from this thread if you don't want to be flooded by comments).
@pantale I'll take over for the actual publication part which is a bit cumbersome. Can you point me to your repository for the source of the paper such that I can make a PR?
