### 2020-06-15 #19 @pantale
@rougier
Sure, I think that everything you want is there:
https://github.com/pantale/ReScience-paper2020
this repository has been updated to the very last version of the paper
