### 2020-06-19 #19 @pantale
@rougier You talk about saving the software onto software heritage. This has already been done, here:
https://archive.softwareheritage.org/browse/directory/e45b8b7efb002231f4081c429c4ca33664d0812d/?origin_url=https://github.com/pantale/ReDynELA
I thing the swid is this:
swh:1:dir:e45b8b7efb002231f4081c429c4ca33664d0812d;origin=https://github.com/pantale/ReDynELA;visit=swh:1:snp:4e9f3ab0d0b2a1e806c310a0aba1ff292dd11338;anchor=swh:1:rev:273db43f7a32f956847ce88c2d0e0b7f432a2bb5;path=/
Olivier
