### 2020-06-19 #19 @rougier
Just sent you a PR. Make sure everything is correct then I will pre-register a DOI and regenerate the paper.
