### 2020-06-19 #19 @pantale
@rougier
I have merged your modifications. Seems that everything is OK.
