### 2020-06-19 #19 @rougier
It's been published ! Congratulations !

See https://rescience.github.io/read/ and https://zenodo.org/record/3901241.
