### 2020-06-19 #19 @pantale
@rougier
Commit merged.
Thanks a lot for your work on the final version.

Olivier
